# Change docker image to production