--- a/docker/build_upload_release.py
+++ b/docker/build_upload_release.py
@@ -26,11 +26,9 @@ if __name__ == "__main__":
     organization = args.organization
 
     # Upload pytorch/torchserve docker binaries
+    try_and_handle(f"./build_image.sh -t {organization}/torchserve:latest", dry_run)
     try_and_handle(
-        f"./build_image.sh -bt dev -t {organization}/torchserve:latest", dry_run
-    )
-    try_and_handle(
-        f"./build_image.sh -bt dev -g -cv cu117 -t {organization}/torchserve:latest-gpu",
+        f"./build_image.sh -g -cv cu117 -t {organization}/torchserve:latest-gpu",
         dry_run,
     )
     try_and_handle(