--- a/docker/docker_nightly.py
+++ b/docker/docker_nightly.py
@@ -32,9 +32,9 @@ if __name__ == "__main__":
     gpu_version = f"{project}:gpu-{get_nightly_version()}"
 
     # Build Nightly images and append the date in the name
-    try_and_handle(f"./build_image.sh -bt dev -t {organization}/{cpu_version}", dry_run)
+    try_and_handle(f"./build_image.sh -t {organization}/{cpu_version}", dry_run)
     try_and_handle(
-        f"./build_image.sh -bt dev -g -cv cu117 -t {organization}/{gpu_version}",
+        f"./build_image.sh -g -cv cu117 -t {organization}/{gpu_version}",
         dry_run,
     )