@msaroufim 2023-04-12
For this image could you please run a few sample inferences on it for an image and a language model? I remember we moved to the dev image in the first place because some stuff was broken for prod just can't remember what exactly