@fabridamicelli 2023-04-14
> For this image could you please run a few sample inferences on it for an image and a language model? I remember we moved to the dev image in the first place because some stuff was broken for prod just can't remember what exactly
> 
> EDIT: I'd also like to see the CUDA image explicitly tested

A side comment if I may: 
For image model (on CPU), [this test in CI](https://github.com/pytorch/serve/blob/9edd461998d2af52b030edc8560493979ae786bc/.github/workflows/test-docker-container.yml#L35) is now covering it.
I just mention it because that might be a good place to extend the test with a language model example (at least for the CPU case).