@agunapal 2023-04-19
@fabridamicelli  I have attached the logs of a test-case with ResNet, HF BERT models.
I did not want to copy paste the code and extend the script.
I think we should extend the testcases in a more organized effort and avoid code duplication. I am looking forward to your design suggestion for how we want to do this.